+++ src/Sensorario/Resources/Schema/Exceptions/NotAllowedPropertyException.php
@@ -1,0 +1,20 @@
+<?php
+
+/**
+ * This file is part of sensorario/resources repository
+ *
+ * (c) Simone Gentili <sensorario@gmail.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Sensorario\Resources\Schema\Exceptions;
+
+use Exception;
+
+final class NotAllowedPropertyException
+    extends Exception
+{
+
+}