+++ src/Sensorario/Resources/Container.php
@@ -5,7 +5,6 @@ namespace Sensorario\Resources;
 use \RuntimeException;
 
 class Container
-    extends ContainerBase
 {
     private $resources;
 
@@ -124,5 +123,45 @@ class Container
     public function globals()
     {
         return $this->globals;
+    }
+
+    public function mandatory($resource)
+    {
+        return $this->getConstraints(
+            'mandatory',
+            $resource
+        );
+    }
+
+    public function defaults($resource)
+    {
+        return $this->getConstraints(
+            'defaults',
+            $resource
+        );
+    }
+
+    public function rules($resource)
+    {
+        return $this->getConstraints(
+            'rules',
+            $resource
+        );
+    }
+
+    public function allowedValues($resource)
+    {
+        return $this->getConstraints(
+            'allowedValues',
+            $resource
+        );
+    }
+
+    public function ranges($resource)
+    {
+        return $this->getConstraints(
+            'allowedRanges',
+            $resource
+        );
     }
 }