+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Release 4.2
 
+ * [close #94] - yaml configuration support
  * [close #69] - documentation improvements
  * [close #70] - property become mandatory when a value or array of value
  * [close #71] - remove duplication and set methods as deprecated